refactor(#1459): decompose tiger_cache.py into cache/tiger/ subpackage (Phase 4)
Split the 3,382-LOC tiger_cache.py monolith into 5 focused modules:
- cache/tiger/bitmap_cache.py (2,212 LOC): CacheKey + TigerCache
- cache/tiger/resource_map.py (369 LOC): TigerResourceMap
- cache/tiger/updater.py (416 LOC): TigerCacheUpdater
- cache/tiger/expander.py (438 LOC): DirectoryGrantExpander

Original tiger_cache.py replaced with backward-compat re-export shim.
All existing imports continue to work unchanged.
